requirements.txt
attrs==18.2.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
# Editable Git install with no remote (dlfinalproject==0.1.0)
-e /media/data/HW/dlfinalproject
entrypoints==0.3
idna==2.8
imageio==2.5.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.9
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterthemes==0.20.0
kiwisolver==1.0.1
lesscpy==0.13.0
MarkupSafe==1.1.1
matplotlib==3.0.2
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.4
numpy==1.16.2
opencv-python==4.1.0.25
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2018.9
pyzmq==18.0.0
qtconsole==4.4.3
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.1
testpath==0.4.2
torch==1.0.1.post2
torchtext==0.3.1
torchvision==0.2.2.post3
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
yapf==0.26.0
